backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
alembic==1.16.4
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
httpx==0.28.1
redis==6.4.0
openai==1.99.9
